Rename feature module to ui module.
