+++ android/app/build.gradle.kts
@@ -26,14 +26,14 @@ android {
 
 dependencies {
   implementation(projects.android.designsystem)
-  implementation(projects.android.feature.discover)
-  implementation(projects.android.feature.library)
-  implementation(projects.android.feature.moreShows)
-  implementation(projects.android.feature.search)
-  implementation(projects.android.feature.seasonDetails)
-  implementation(projects.android.feature.settings)
-  implementation(projects.android.feature.showDetails)
-  implementation(projects.android.feature.trailers)
+  implementation(projects.android.ui.discover)
+  implementation(projects.android.ui.library)
+  implementation(projects.android.ui.moreShows)
+  implementation(projects.android.ui.search)
+  implementation(projects.android.ui.seasonDetails)
+  implementation(projects.android.ui.settings)
+  implementation(projects.android.ui.showDetails)
+  implementation(projects.android.ui.trailers)
   implementation(projects.shared)
   implementation(projects.core.base)
   implementation(projects.core.util)
