+++ android/ui/discover/build.gradle.kts
@@ -1,0 +1,17 @@
+plugins { id("plugin.tvmaniac.compose.library") }
+
+android { namespace = "com.thomaskioko.tvmaniac.ui.discover" }
+
+dependencies {
+  api(projects.presentation.discover)
+
+  implementation(projects.android.designsystem)
+  implementation(projects.android.resources)
+
+  implementation(libs.androidx.compose.foundation)
+  implementation(libs.androidx.compose.material3)
+  implementation(libs.androidx.compose.ui.util)
+  implementation(libs.androidx.compose.runtime)
+  implementation(libs.decompose.extensions.compose)
+  implementation(libs.snapper)
+}
