+++ android/ui/library/build.gradle.kts
@@ -1,0 +1,16 @@
+plugins { id("plugin.tvmaniac.compose.library") }
+
+android { namespace = "com.thomaskioko.tvmaniac.ui.library" }
+
+dependencies {
+  api(projects.presentation.library)
+
+  implementation(projects.android.designsystem)
+  implementation(projects.android.resources)
+
+  implementation(libs.androidx.compose.foundation)
+  implementation(libs.androidx.compose.material3)
+  implementation(libs.androidx.compose.runtime)
+  implementation(libs.decompose.extensions.compose)
+  implementation(libs.kotlinx.collections)
+}
