+++ android/ui/more-shows/.gitignore
