+++ android/ui/search/.gitignore
