+++ android/ui/settings/.gitignore
