+++ android/ui/show-details/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins { id("plugin.tvmaniac.compose.library") }
+
+android { namespace = "com.thomaskioko.tvmaniac.ui.showdetails" }
+
+dependencies {
+  api(projects.presentation.showDetails)
+
+  implementation(projects.android.designsystem)
+  implementation(projects.android.resources)
+
+  implementation(libs.androidx.compose.constraintlayout)
+  implementation(libs.androidx.compose.foundation)
+  implementation(libs.androidx.compose.material.icons)
+  implementation(libs.androidx.compose.material3)
+  implementation(libs.androidx.compose.runtime)
+  implementation(libs.decompose.extensions.compose)
+  implementation(libs.snapper)
+}
