+++ settings.gradle.kts
@@ -23,14 +23,14 @@ rootProject.name = "tv-maniac"
 include(
   ":android:app",
   ":android:designsystem",
-  ":android:feature:discover",
-  ":android:feature:library",
-  ":android:feature:more-shows",
-  ":android:feature:search",
-  ":android:feature:season-details",
-  ":android:feature:settings",
-  ":android:feature:show-details",
-  ":android:feature:trailers",
+  ":android:ui:discover",
+  ":android:ui:library",
+  ":android:ui:more-shows",
+  ":android:ui:search",
+  ":android:ui:season-details",
+  ":android:ui:settings",
+  ":android:ui:show-details",
+  ":android:ui:trailers",
   ":android:resources",
   ":core:base",
   ":core:logger",
